[Issue #134] Renable static rendering (#85)
